pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "mixtape"
version = "0.0.0"
requires-python = ">=3.11"
license = "Apache-2.0"
license-files = ["LICENSE", "NOTICE"]
classifiers = ["Private :: Do Not Upload"]
dependencies = [
  # Runtime dependencies, always needed
  "ale-py",
  "autorom[accept-rom-license]",
  "celery",
  "django[argon2]",
  "django-allauth",
  "django-auth-style",
  "django-click",
  "django-cors-headers",
  "django-environ",
  "django-extensions",
  "django-filter",
  "django-oauth-toolkit",
  "django-resonant-settings[allauth,celery]",
  "django-resonant-utils",
  "django-stubs-ext",
  "djangorestframework",
  "drf-yasg",
  "einops",
  "gymnasium[atari, other]",
  # PettingZoo[butterfly] extra pulls in both pymunk and pygame, but the pypi version of PettingZoo
  # is extremely outdated and depends on a version of pymunk that does not build and has no bdist
  # wheel for our target platforms. So we manually specify those transitive dependency versions.
  "pymunk==6.11.0",
  "pygame>=2.3.0",
  "PettingZoo>=1.24.3",
  "Pillow",
  "psycopg[binary]",
  "ray[default, rllib]==2.34.0",
  "rich",
  "scikit-learn",
  "SuperSuit",
  "torch",
  "umap-learn",
  "whitenoise[brotli]",
  # Production-only
  "django-s3-file-field[s3]",
  "django-storages[s3]",
  "gunicorn",
  "sentry-sdk[celery,django,pure_eval]",
]

[project.optional-dependencies]
development = [
  # Runtime dependencies, only needed with "development" settings
  "django-browser-reload",
  "django-debug-toolbar",
  "django-minio-storage",
  "django-s3-file-field[minio]",
  "ipython",
  "watchdog",
  "werkzeug",
]

[dependency-groups]
dev = [
  # Additional developer tools
  # The "dev" dependency group is installed by default,
  # so use this to install "development" extras by default too
  "mixtape[development]",
  "tox",
  "tox-uv",
  "django-minio-storage",
  "django-s3-file-field[minio]",
]
lint = [
  "flake8",
  "flake8-black",
  "flake8-bugbear",
  "flake8-docstrings",
  "flake8-isort",
  "flake8-quotes",
  "pep8-naming",
]
format = [
  "black",
  "isort",
]
type = [
  "mypy",
  "celery-types",
  "django-stubs[compatible-mypy]",
  "djangorestframework-stubs",
]
test = [
  "factory-boy",
  "pytest",
  "pytest-django",
  "pytest-mock",
  # Enable with "pytest --random-order ..."
  "pytest-random-order",
  # Enable with "pytest --count=... ..."
  "pytest-repeat",
]

[tool.hatch.build]
packages = [
  "mixtape",
]

[tool.uv]
constraint-dependencies = [
  # Earlier versions of "numba" (required by "umap-learn") didn't specify an upper bound for
  # its compatability with "numpy", which led to incorrect versions of "numpy" and "umap-learn".
  "numba >= 0.61.0",
]

[tool.black]
line-length = 100
skip-string-normalization = true
target-version = ["py311"]

[tool.isort]
profile = "black"
line_length = 100
# Sort by name, don't cluster "from" vs "import"
force_sort_within_sections = true
# Combines "as" imports on the same line
combine_as_imports = true

[tool.mypy]
files = [
  "mixtape",
]
check_untyped_defs = true
ignore_missing_imports = true
show_error_codes = true
warn_redundant_casts = true
warn_unused_configs = true
warn_unused_ignores = true
plugins = [
    "mypy_django_plugin.main",
]

[tool.django-stubs]
django_settings_module = "mixtape.settings.testing"

[tool.pytest.ini_options]
addopts = [
  # Test utilities should be imported absolutely from the pythonpath,
  # so use the Pytest-recommended "importlib" mode
  "--import-mode=importlib",
  "--strict-config",
  "--strict-markers",
  "--showlocals",
  "--verbose",
  # Specifying as "--ds" will override any value in the environment
  "--ds=mixtape.settings.testing",
]
filterwarnings = [
  "error",
  # pytest often causes unclosed socket warnings
  "ignore:unclosed <socket\\.socket:ResourceWarning",
]
